Move bundlers helpers in their own file
