+++ packages/core/src/helpers.ts
@@ -4,10 +4,7 @@
 
 import { INJECTED_FILE } from '@dd/core/constants';
 import { isInjectionFile } from '@dd/core/helpers/plugins';
-import type { BundlerFullName, GlobalContext, Logger, ResolvedEntry } from '@dd/core/types';
-import type { PluginBuild } from 'esbuild';
 import fs from 'fs';
-import { glob } from 'glob';
 import path from 'path';
 
 // Format a duration 0h 0m 0s 0ms
@@ -25,74 +22,6 @@ export const formatDuration = (duration: number) => {
         }`.trim();
     // Split here so we can show 0ms in case we have a duration of 0.
     return `${timeString}${!timeString || milliseconds ? ` ${milliseconds}ms` : ''}`.trim();
-};
-
-// https://esbuild.github.io/api/#glob-style-entry-points
-const getAllEntryFiles = (filepath: string): string[] => {
-    if (!filepath.includes('*')) {
-        return [filepath];
-    }
-
-    const files = glob.sync(filepath);
-    return files;
-};
-
-// Parse, resolve and return all the entries of esbuild.
-export const getEsbuildEntries = async (
-    build: PluginBuild,
-    context: GlobalContext,
-    log: Logger,
-): Promise<ResolvedEntry[]> => {
-    const entries: { name?: string; resolved: string; original: string }[] = [];
-    const entryPoints = build.initialOptions.entryPoints;
-    const entryPaths: { name?: string; path: string }[] = [];
-    const resolutionErrors: string[] = [];
-
-    if (Array.isArray(entryPoints)) {
-        for (const entry of entryPoints) {
-            const fullPath = entry && typeof entry === 'object' ? entry.in : entry;
-            entryPaths.push({ path: fullPath });
-        }
-    } else if (entryPoints && typeof entryPoints === 'object') {
-        entryPaths.push(
-            ...Object.entries(entryPoints).map(([name, filepath]) => ({ name, path: filepath })),
-        );
-    }
-
-    // Resolve all the paths.
-    const proms = entryPaths
-        .flatMap((entry) =>
-            getAllEntryFiles(entry.path).map<[{ name?: string; path: string }, string]>((p) => [
-                entry,
-                p,
-            ]),
-        )
-        .map(async ([entry, p]) => {
-            const result = await build.resolve(p, {
-                kind: 'entry-point',
-                resolveDir: context.cwd,
-            });
-
-            if (result.errors.length) {
-                resolutionErrors.push(...result.errors.map((e) => e.text));
-            }
-
-            if (result.path) {
-                // Store them for later use.
-                entries.push({
-                    name: entry.name,
-                    resolved: result.path,
-                    original: entry.path,
-                });
-            }
-        });
-
-    for (const resolutionError of resolutionErrors) {
-        log.error(resolutionError);
-    }
-
-    await Promise.all(proms);
-    return entries;
 };
 
 // Truncate a string to a certain length.
@@ -116,10 +45,6 @@ export const truncateString = (
 
     return `${str.slice(0, leftStop)}${placeholder}${str.slice(-rightStop)}`;
 };
-
-// From a bundler's name, is it part of the "xpack" family?
-export const isXpack = (bundlerName: BundlerFullName) =>
-    ['rspack', 'webpack4', 'webpack5', 'webpack'].includes(bundlerName);
 
 let index = 0;
 export const getUniqueId = () => `${Date.now()}.${performance.now()}.${++index}`;
