+++ packages/plugins/injection/src/index.ts
@@ -2,8 +2,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
+import { isXpack } from '@dd/core/helpers/bundlers';
 import { isInjectionFile } from '@dd/core/helpers/plugins';
-import { getUniqueId, isXpack } from '@dd/core/helpers';
+import { getUniqueId } from '@dd/core/helpers';
 import {
     InjectPosition,
     type GlobalContext,
